dom/imptests/webapps/DOMCore/tests/submissions/Ms2ger/test_Event-constructors.html

<!doctype html>
<title>Event constructors</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id=log></div>
<script>
test(function() {
  var ev = new Event("test")
  assert_equals(ev.type, "test")
  assert_equals(ev.target, null)
  assert_equals(ev.currentTarget, null)
  assert_equals(ev.eventPhase, Event.AT_TARGET)
  assert_equals(ev.bubbles, false)
  assert_equals(ev.cancelable, false)
  assert_equals(ev.defaultPrevented, false)
  assert_equals(ev.isTrusted, false)
  assert_true(ev.timeStamp > 0)
  assert_true("initEvent" in ev)
})
test(function() {
  var ev = new Event("I am an event", { bubbles: true, cancelable: false})
  assert_equals(ev.type, "I am an event")
  assert_equals(ev.bubbles, true)
  assert_equals(ev.cancelable, false)
})
test(function() {
  var ev = new Event("@", { bubblesIGNORED: true, cancelable: true})
  assert_equals(ev.type, "@")
  assert_equals(ev.bubbles, false)
  assert_equals(ev.cancelable, true)
})
test(function() {
  var ev = new Event("Xx", { cancelable: true})
  assert_equals(ev.type, "Xx")
  assert_equals(ev.bubbles, false)
  assert_equals(ev.cancelable, true)
})
test(function() {
  var ev = new Event("Xx", {})
  assert_equals(ev.type, "Xx")
  assert_equals(ev.bubbles, false)
  assert_equals(ev.cancelable, false)
})
test(function() {
  var ev = new Event("Xx", {bubbles: true, cancelable: false, sweet: "x"})
  assert_equals(ev.type, "Xx")
  assert_equals(ev.bubbles, true)
  assert_equals(ev.cancelable, false)
  assert_equals(ev.sweet, undefined)
})
test(function() {
  var ev = new CustomEvent("$", {detail: 54, sweet: "x", sweet2: "x", cancelable:true})
  assert_equals(ev.type, "$")
  assert_equals(ev.bubbles, false)
  assert_equals(ev.cancelable, true)
  assert_equals(ev.sweet, undefined)
  assert_equals(ev.detail, 54)
})
</script>
